+++ scilab/modules/ast/includes/types/cell.hxx
@@ -109,6 +109,8 @@ public :
 
     virtual bool transpose(InternalType *& out);
 
+    bool getMemory(int* _piSize, int* _piSizePlusType);
+
 private :
     virtual InternalType*   getNullValue();
     virtual Cell*           createEmpty(int _iDims, int* _piDims, bool _bComplex = false);