+++ scilab/modules/ast/includes/types/macro.hxx
@@ -66,6 +66,8 @@ public :
 
     ast::SeqExp*                getBody();
 
+    bool                        getMemory(int* _piSize, int* _piSizePlusType);
+
     /* return type as string ( double, int, cell, list, ... )*/
     virtual std::wstring        getTypeStr() const
     {