+++ scilab/modules/ast/includes/types/singlestruct.hxx
@@ -85,6 +85,8 @@ public :
         return L"sst";
     }
 
+    bool                                    getMemory(int* _piSize, int* _piSizePlusType);
+
 private :
 
     std::unordered_map<std::wstring, int> m_wstFields;