+++ scilab/modules/ast/includes/types/sparse.hxx
@@ -80,6 +80,8 @@ struct EXTERN_AST Sparse : GenericType
         return true;
     }
     void finalize();
+    
+    bool getMemory(int *_piSize, int* _piSizePlusType);
 
     /*data management member function defined for compatibility with the Double API*/
     Sparse* set(int _iRows, int _iCols, double _dblReal, bool _bFinalize = true);
@@ -515,6 +517,8 @@ struct EXTERN_AST SparseBool : GenericType
         return true;
     }
     void finalize();
+
+    bool getMemory(int *_piSize, int* _piSizePlusType);
 
     bool toString(std::wostringstream& ostr);