+++ scilab/modules/ast/includes/types/tlist.hxx
@@ -76,6 +76,8 @@ public :
 
     bool                            toString(std::wostringstream& ostr);
 
+    bool                            getMemory(int* _piSize, int* _piSizePlusType);
+
 private :
 };
 }